Resizing horizontal simple screen elements
Also janpom, WHY WOULD YOU SET THE X AND Y COORIDNATES TO BE UINT_8T???

+++ tft_util.h
@@ -23,11 +23,11 @@
 #include <Adafruit_ILI9341.h>
 
 void tft_util_draw_digit(
-        Adafruit_ILI9341* tft, uint8_t digit, uint8_t x, uint8_t y,
+        Adafruit_ILI9341* tft, uint8_t digit, uint16_t x, uint16_t y,
         uint16_t fg_color, uint16_t bg_color, uint8_t magnify = 1);
 
 void tft_util_draw_number(
-        Adafruit_ILI9341* tft, char *number, uint8_t x, uint8_t y,
+        Adafruit_ILI9341* tft, char *number, uint16_t x, uint16_t y,
         uint16_t fg_color, uint16_t bg_color, uint8_t spacing, uint8_t magnify = 1);
 
 uint16_t progress_to_color(float progress, Adafruit_ILI9341* tft);
